implement omp on demand tls

+++ include/oneapi/dpl/pstl/parallel_backend_serial.h
@@ -50,10 +50,10 @@ struct __thread_enumerable_storage
     {
     }
 
-    std::size_t
+    std::uint32_t
     size() const
     {
-        return std::size_t{1};
+        return std::uint32_t{1};
     }
 
     _StorageType&
@@ -63,7 +63,7 @@ struct __thread_enumerable_storage
     }
 
     _StorageType&
-    get_with_id(std::size_t __i)
+    get_with_id(std::uint32_t __i)
     {
         return get();
     }
